Use the index data structures to speed up the composite aggregation when possible
The composite aggregation does not need to collect all documents to retrieve the next N sorted buckets.
For instance if the query is a match all we can iterate the terms dictionary and the postings lists to retrieve
the next N sorted terms and their associated documents instead of collecting all documents that a normal aggregation would do.
For numerics the bkd tree can be used similarly to restrict the collection to the next N sorted values and their documents.
This change implements this idea by using the first source that is set on the composite aggregation to lead the iterations.
It works for keywords, longs, integers and dates and is activated under certain conditions:
 * The field must be indexed.
 * The sort order of the primary source must be ascending.
 * The query must be a match all query or a range query on the source field.
When activated this option terminates the query collection for each segment and uses the index data structures to restrict the search
to the next N sorted buckets of the primary source. This ensures that the selected documents will contain at least the N next sorted composite buckets.
It also works for date_histogram since the rounding does not modify the sort order but it is deactivated for range queries that uses
non-rounded boundaries.
Tests showed that this optimization can speed up composite aggregations by a big factor since it allows the aggregation to not collect
all documents on each page. For instance a simple term composite aggregation only needs to checks the postings of the next N sorted terms,
N being the requested number of buckets on each page.