Remove ParsedQuery and handle named queries in every query context
ParsedQuery is used to keep the Lucene Query and their associated query names.
Now that the SearchExecutionContext keeps a copy of the named queries, we can remove
this class as it contains only redundant informations.
As a side effect, this change fixes the named queries that are set in the rescore query.

Closes #65563